Refactor to be on Api::V8::Core module
