+++ app/controllers/api/v8/apidocs_controller.rb
@@ -130,11 +130,11 @@ module Api
         Api::V8::Organizations::Courses::Exercises::Users::PointsController,
         Api::V8::Organizations::Courses::SubmissionsController,
         Api::V8::Organizations::Courses::Users::SubmissionsController,
-        Api::V8::ExercisesController,
-        Api::V8::Exercises::SolutionsController,
         Api::V8::SubmissionsController,
         Api::V8::UsersController,
         Api::V8::Core::Organizations::Courses::UnlocksController,
+        Api::V8::Core::ExercisesController,
+        Api::V8::Core::Exercises::SolutionsController,
         Course,
         Exercise,
         AvailablePoint,
